[IMP] base: Changed modifier syntax (invisible, required, readonly)
Goal:
* Simplified modifiers to only have one way to define modifiers (remove
`attrs` and `states` and replaced by corresponding attribute);
* Use domain in attributes `required`, `readonly`, `invisible`;
* More accurate validation of xml views.

This commit converts the old syntax (views get coresponding domains)
and adds a warning only in dev mode. The conversion of the xml files
will be done in future commits then the warning will be applied all the
time.

Use:
The `required`, `readonly` and `invisible` attributes can only have
values of `True`, `False`, 1, 0 or a domain. The domains can contains
contextual value and will be evaluate by the javascript.

```xml
    <field name="field_a" readonly="not context.get('show_a', True)" attrs="{'readonly': [('field_b', '>', 0), ('field_c', '=', parent.c)]}"/>
    <field name="field_b" states="draft"/>
```
will be replaced by
```xml
    <field name="field_a" readonly="['|', not context.get('show_a'), '&', ('field_b', '>', 0), ('field_c', '=', parent.c)]"/>
    <field name="field_b" invisible="[('state', '!=', 'draft')]"/>
```

Details:
The parser will normalize the different domains/attributes including
the domains that cannot be evaluated literally (with contextual values).
Thus it is possible to combine (with AND and OR) fields such as:
`[('field_a', '=', 'value')]` and `[('field_b', '=', context.get( 'b'))]`
and `[('field_c', '=', field_other)]`.

Validating the view will raise an error if the `field_a`, `field_b` or
`field_c` is missing on the model, or if the field `field_other` is
missing in the xml view.

(nb: `domain` attribute on field are also validated)

opw-2495504